move detokenizer into engine
